Consolidate CodeResourceDefinition classes
